emolog, emolog_embedded: projects now use relative Tivaware folder.
Tivaware project is now also in Git (separate repo) and should be placed side by side with the Emolog repo. Accordingly, the machine-specific SW_ROOT definition in CCS options can (must!) be removed. It is now defined in each project (to a relative value as noted above).

Also brought Debug configuration up to speed (was neglected recently in favor of Release, and had several problems).