emolog_pc: use max_samples, not max_ticks
Since the embedded side may restart during a single recording for
the long term use case, don't actually count ticks, instead count
the received samples.

For the limited samples use case, i.e. recording a few seconds, this
is not really an issue. But it makes it easier to test the restarting
scenario, and may be useful (i.e. running 1 hour of samples)