0.2.7 with new tools
no other changes. emotool itself has not been changed. added tools
include:

emotool-vars
emotool-dwarf
emotool-dwarf-dump

The first two are for external users to reuse the logic emotool uses
for reading the vars and the ELF file, but with much less overhead.

The last is a straight debugging tool, but can also be used by
an external UI.