emotool: add --check-timestamp
Half of the code is for testing. There is new logic, defaulting to
off and activated by passing the new '--check-timestamp' to emolog
(will be turned to default on in the future).

The logic is:

look for a variable in ELF file by special name
(consts.BUILD_TIMESTAMP_VARNAME) currently set to build_timestamp

If not found in ELF, exit.

If found, query target for it's value.

If value is not the same, exit (printing older/newer).